Wrote in Day 0
It's still missing some stuff (sound effects), but that can be implemented later. [this is a reminder for me]
I will need an " ! " emote, also need to check if we have a fade out option. And what position is the bonsai? If we can't fade out Kodama, can we hide it behind the bonsai? 
 I've added some cards for human shadows and water droplets.